requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-modeltranslation==0.18.11
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
idna==3.6
Pillow==10.1.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
